requirements.txt
pandas
requests
numpy
matplotlib
openpyxl
python-dotenv